:fire: Removed unused files Also moved around types / story components so they weren't affecting code coverage
