+++ src/components/DataDisplay/Tutorial/stories/TutorialStoryComponent.tsx
@@ -3,8 +3,8 @@ import { FC } from 'react';
 import { Button, Checkbox, TextField } from '@equinor/eds-core-react';
 import { tokens } from '@equinor/eds-tokens';
 
-import HighlightBlocks from './HighlightBlocks/HighlightBlocks';
-import Tutorial, { TutorialProps } from './Tutorial';
+import HighlightBlocks from '../HighlightBlocks/HighlightBlocks';
+import Tutorial, { TutorialProps } from '../Tutorial';
 import { useTutorialSteps } from 'src/providers/TutorialStepsProvider';
 
 import styled from 'styled-components';
