+++ src/components/Icons/ApplicationIcon/FourDInsight.tsx
@@ -1,8 +1,8 @@
 import { forwardRef } from 'react';
 
-import { AppIconProps } from '../index';
 import ApplicationIconBase, { ShapeProps } from './ApplicationIconBase';
 import { fourDInsight } from './ApplicationIconCollection';
+import { AppIconProps } from 'src/types/AppIcon';
 
 const shapes: ShapeProps[] = [
   {
