+++ src/components/Icons/FeedBackIcon/FeedBackIcon.tsx
@@ -1,10 +1,10 @@
 import { FC } from 'react';
 
-import { AppIconProps } from '../index';
 import NegativeFilled from './NegativeFilled';
 import NegativeOutlined from './NegativeOutlined';
 import PositiveFilled from './PositiveFilled';
 import PositiveOutlined from './PositiveOutlined';
+import { AppIconProps } from 'src/types/AppIcon';
 
 interface FeedBackIconData {
   iconName: string;
